# api/openapi-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Envrouter API
  description: Envrouter API
  version: 1.0.0
servers:
  - url: http://localhost:8080/
components:
  schemas:
...
    RefBinding:
      type: object
      required:
        - environment
        - application
        - ref
      properties:
        environment:
          type: string
        application:
          type: string
        ref:
          type: string
...
paths:
...
  "/api/v1/refBindings":
    get:
      parameters:
        - in: query
          name: application
          schema:
            type: string
        - in: query
          name: environment
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        "200":
          description: Get list of RefBinding
          content:
            "application/json":
              "schema":
                "type": "array"
                "items":
                  "$ref": "#/components/schemas/RefBinding"
    post:
      description: Create RefBinding
      requestBody:
        content:
          "application/json":
            "schema":
              "$ref": "#/components/schemas/RefBinding"
      responses:
        "201":
          description: Success
          content:
            "application/json":
              "schema":
                "$ref": "#/components/schemas/RefBinding"
...