# internal/envrouter/auth/auth.go
// Package auth provides optional OIDC authentication (Keycloak, Authelia, or
// any discovery-compliant provider). It is enabled by setting
// ENVROUTER_OIDC_ISSUER; without it every request passes through with an
// anonymous actor. Sessions are cookie-based (the raw ID token in an
// HttpOnly cookie) because EventSource cannot send Authorization headers.
//
// When the provider issues a refresh token (Keycloak by default; providers
// like Authelia require "offline_access" in ENVROUTER_OIDC_SCOPES), it is
// stored in a second HttpOnly cookie and used to silently mint a fresh ID
// token as the current one nears expiry, keeping the session alive without a
// re-login round-trip. Cookie Max-Age tracks the real token lifetimes.
package auth

import (
	"context"
	"crypto/rand"
	"crypto/subtle"
	"encoding/base64"
	"errors"
	"fmt"
	"net/http"
	"net/url"
	"os"
	"strings"
	"time"

	"github.com/coreos/go-oidc/v3/oidc"
	"github.com/gin-gonic/gin"
	log "github.com/sirupsen/logrus"
	"golang.org/x/oauth2"
)

const (
	sessionCookie = "envrouter_session"
	refreshCookie = "envrouter_refresh"
	stateCookie   = "envrouter_oauth_state"
	actorKey      = "envrouter_actor"

	// refreshSkew: proactively refresh once less than this remains on the ID
	// token. ponytail: coupled to the frontend's 5-minute userinfo poll — one
	// poll always lands inside this window, so the token refreshes before it
	// expires. Bump both together if the poll interval changes.
	refreshSkew = 6 * time.Minute
)

// Actor identifies who performs an API request.
type Actor struct {
	UserIdentifier string
	FullName       string
	Email          string
	IP             string
	Groups         []string
	// Admin marks a request authenticated via the static API token — it is
	// permitted at every level regardless of OIDC group config.
	Admin bool
}

type Config struct {
	Issuer         string
	ClientID       string
	ClientSecret   string
	RedirectURL    string
	Scopes         []string
	InsecureCookie bool
	// Claim mapping: fallback lists of OIDC claim names, first non-empty wins.
	ClaimUserID   []string
	ClaimFullName []string
	ClaimEmail    []string
	ClaimGroups   []string
	// Group required at each authorization level; empty = allow all
	// authenticated users at that level.
	GroupView      string
	GroupDeploy    string
	GroupConfigure string
	// APIToken, when set, enables full-access machine auth via
	// `Authorization: Bearer <token>` (works whether or not OIDC is enabled).
	APIToken string
	// AnonymousLegacyReads, when true, lets allow-listed legacy v1 GET
	// endpoints be read without a session even when OIDC is enabled — for
	// CI/automation that predates auth.
	AnonymousLegacyReads bool
}

func getEnv(key, fallback string) string {
	if v := os.Getenv(key); v != "" {
		return v
	}
	return fallback
}

func splitClaims(s string) []string {
	var result []string
	for _, part := range strings.Split(s, ",") {
		if trimmed := strings.TrimSpace(part); trimmed != "" {
			result = append(result, trimmed)
		}
	}
	return result
}

func ConfigFromEnv() Config {
	scopes := strings.Fields(strings.ReplaceAll(getEnv("ENVROUTER_OIDC_SCOPES", "openid profile email"), ",", " "))
	return Config{
		Issuer:         os.Getenv("ENVROUTER_OIDC_ISSUER"),
		ClientID:       os.Getenv("ENVROUTER_OIDC_CLIENT_ID"),
		ClientSecret:   os.Getenv("ENVROUTER_OIDC_CLIENT_SECRET"),
		RedirectURL:    os.Getenv("ENVROUTER_OIDC_REDIRECT_URL"),
		Scopes:         scopes,
		InsecureCookie: os.Getenv("ENVROUTER_OIDC_INSECURE_COOKIE") == "true",
		// defaults work with Keycloak and Authelia out of the box; sub is
		// always present per the OIDC spec, so the required field can't
		// end up empty with default mapping
		ClaimUserID:    splitClaims(getEnv("ENVROUTER_OIDC_CLAIM_USER_IDENTIFIER", "preferred_username,sub")),
		ClaimFullName:  splitClaims(getEnv("ENVROUTER_OIDC_CLAIM_FULLNAME", "name")),
		ClaimEmail:     splitClaims(getEnv("ENVROUTER_OIDC_CLAIM_EMAIL", "email")),
		ClaimGroups:    splitClaims(getEnv("ENVROUTER_OIDC_CLAIM_GROUPS", "groups")),
		GroupView:      os.Getenv("ENVROUTER_OIDC_GROUP_VIEW"),
		GroupDeploy:    os.Getenv("ENVROUTER_OIDC_GROUP_DEPLOY"),
		GroupConfigure: os.Getenv("ENVROUTER_OIDC_GROUP_CONFIGURE"),

		APIToken:             os.Getenv("ENVROUTER_API_TOKEN"),
		AnonymousLegacyReads: os.Getenv("ENVROUTER_ANONYMOUS_LEGACY_READS") == "true",
	}
}

type Service struct {
	cfg      Config
	provider *oidc.Provider
	verifier *oidc.IDTokenVerifier
	oauth    oauth2.Config
}

func New(ctx context.Context, cfg Config) (*Service, error) {
	if cfg.APIToken != "" {
		log.Info("Static API token auth enabled (Authorization: Bearer)")
	}
	if cfg.AnonymousLegacyReads {
		log.Info("Anonymous legacy v1 read endpoints enabled")
	}
	if cfg.Issuer == "" {
		log.Info("OIDC authentication disabled (ENVROUTER_OIDC_ISSUER not set)")
		return &Service{cfg: cfg}, nil
	}
	if cfg.ClientID == "" || cfg.RedirectURL == "" {
		return nil, errors.New("ENVROUTER_OIDC_CLIENT_ID and ENVROUTER_OIDC_REDIRECT_URL are required when ENVROUTER_OIDC_ISSUER is set")
	}
	provider, err := oidc.NewProvider(ctx, cfg.Issuer)
	if err != nil {
		return nil, fmt.Errorf("OIDC discovery for %s failed: %w", cfg.Issuer, err)
	}
	log.Infof("OIDC authentication enabled, issuer %s", cfg.Issuer)
	return &Service{
		cfg:      cfg,
		provider: provider,
		verifier: provider.Verifier(&oidc.Config{ClientID: cfg.ClientID}),
		oauth: oauth2.Config{
			ClientID:     cfg.ClientID,
			ClientSecret: cfg.ClientSecret,
			Endpoint:     provider.Endpoint(),
			RedirectURL:  cfg.RedirectURL,
			Scopes:       cfg.Scopes,
		},
	}, nil
}

func (s *Service) Enabled() bool {
	return s.provider != nil
}

// permitted reports whether the actor may act at a level whose required group
// is `group`. Auth disabled ⇒ always allowed; empty group ⇒ any authenticated
// user; otherwise the actor must be a member.
func (s *Service) permitted(a Actor, group string) bool {
	if !s.Enabled() {
		return true
	}
	if a.Admin {
		return true
	}
	if group == "" {
		return true
	}
	return containsString(a.Groups, group)
}

// Levels are hierarchical: configure ⊃ deploy ⊃ view, so an admin needs only
// the configure group, an editor only the deploy group.
func (s *Service) CanView(a Actor) bool { return s.permitted(a, s.cfg.GroupView) || s.CanDeploy(a) }
func (s *Service) CanDeploy(a Actor) bool {
	return s.permitted(a, s.cfg.GroupDeploy) || s.CanConfigure(a)
}
func (s *Service) CanConfigure(a Actor) bool { return s.permitted(a, s.cfg.GroupConfigure) }

// safeRedirect sanitizes the `rd` param to a same-site absolute path, defeating
// open-redirect tricks including backslash normalization. Query string is
// dropped.
func safeRedirect(rd string) string {
	if rd == "" || strings.ContainsAny(rd, "\\") {
		return "/"
	}
	u, err := url.Parse(rd)
	if err != nil || u.IsAbs() || u.Host != "" || !strings.HasPrefix(u.Path, "/") || strings.HasPrefix(u.Path, "//") || strings.HasPrefix(u.Path, "/auth/") {
		return "/"
	}
	return u.Path
}

func resolveClaim(claims map[string]interface{}, fallbacks []string) string {
	for _, name := range fallbacks {
		if v, ok := claims[name].(string); ok && v != "" {
			return v
		}
	}
	return ""
}

// resolveClaimSlice walks the fallback list and returns the first claim whose
// value is a non-empty list of strings (OIDC claims arrive as []interface{}).
func resolveClaimSlice(claims map[string]interface{}, fallbacks []string) []string {
	for _, name := range fallbacks {
		raw, ok := claims[name].([]interface{})
		if !ok {
			continue
		}
		var result []string
		for _, e := range raw {
			if s, ok := e.(string); ok && s != "" {
				result = append(result, s)
			}
		}
		if len(result) > 0 {
			return result
		}
	}
	return nil
}

func containsString(list []string, v string) bool {
	for _, e := range list {
		if e == v {
			return true
		}
	}
	return false
}

func (s *Service) actorFromIDToken(ctx context.Context, rawIDToken string) (Actor, time.Time, error) {
	idToken, err := s.verifier.Verify(ctx, rawIDToken)
	if err != nil {
		return Actor{}, time.Time{}, err
	}
	claims := map[string]interface{}{}
	if err := idToken.Claims(&claims); err != nil {
		return Actor{}, time.Time{}, err
	}
	actor := Actor{
		UserIdentifier: resolveClaim(claims, s.cfg.ClaimUserID),
		FullName:       resolveClaim(claims, s.cfg.ClaimFullName),
		Email:          resolveClaim(claims, s.cfg.ClaimEmail),
		Groups:         resolveClaimSlice(claims, s.cfg.ClaimGroups),
	}
	if actor.UserIdentifier == "" {
		return Actor{}, time.Time{}, fmt.Errorf("none of the claims %v yielded a user identifier", s.cfg.ClaimUserID)
	}
	return actor, idToken.Expiry, nil
}

// setSessionCookies writes the session (ID token) and, when present, refresh
// cookies with Max-Age matching the real token lifetimes: the browser drops
// each cookie exactly when its token expires.
func (s *Service) setSessionCookies(c *gin.Context, rawIDToken string, idExpiry time.Time, token *oauth2.Token) {
	secure := !s.cfg.InsecureCookie
	c.SetSameSite(http.SameSiteLaxMode)
	c.SetCookie(sessionCookie, rawIDToken, int(time.Until(idExpiry).Seconds()), "/", "", secure, true)
	if token != nil && token.RefreshToken != "" {
		// OIDC does not expose refresh-token expiry in a standard claim;
		// Keycloak sends refresh_expires_in. Absent that, use a browser-session
		// cookie (Max-Age 0) so it lives until the browser closes.
		maxAge := 0
		if v, ok := token.Extra("refresh_expires_in").(float64); ok && v > 0 {
			maxAge = int(v)
		}
		c.SetCookie(refreshCookie, token.RefreshToken, maxAge, "/", "", secure, true)
	}
}

// clearSessionCookies removes both auth cookies (used on logout and on a failed
// mandatory refresh).
func (s *Service) clearSessionCookies(c *gin.Context) {
	secure := !s.cfg.InsecureCookie
	c.SetSameSite(http.SameSiteLaxMode)
	c.SetCookie(sessionCookie, "", -1, "/", "", secure, true)
	c.SetCookie(refreshCookie, "", -1, "/", "", secure, true)
}

// refresh redeems the refresh token for a new ID token, verifies it, and
// re-issues both cookies. Returns the new actor on success.
func (s *Service) refresh(c *gin.Context, refreshToken string) (Actor, error) {
	token, err := s.oauth.TokenSource(c.Request.Context(), &oauth2.Token{RefreshToken: refreshToken}).Token()
	if err != nil {
		return Actor{}, fmt.Errorf("refresh token exchange failed: %w", err)
	}
	rawIDToken, ok := token.Extra("id_token").(string)
	if !ok {
		return Actor{}, errors.New("refresh response contained no id_token")
	}
	actor, idExpiry, err := s.actorFromIDToken(c.Request.Context(), rawIDToken)
	if err != nil {
		return Actor{}, fmt.Errorf("refreshed id_token rejected: %w", err)
	}
	s.setSessionCookies(c, rawIDToken, idExpiry, token)
	return actor, nil
}

// actorFromSession resolves the request actor from the session cookie,
// refreshing the ID token when it is expired or close to expiry. On a
// mandatory-but-failed refresh it clears the cookies and returns the error so
// callers fall through to the 401 path.
//
// ponytail: concurrent tabs may race on refresh-token rotation — the loser gets
// a 401 and the SPA re-logs in (usually silent via the IdP's SSO cookie).
// Upgrade path: single-flight per session if that ever becomes noticeable.
func (s *Service) actorFromSession(c *gin.Context) (Actor, error) {
	rawIDToken, cookieErr := c.Cookie(sessionCookie)
	if cookieErr == nil {
		actor, idExpiry, err := s.actorFromIDToken(c.Request.Context(), rawIDToken)
		if err == nil {
			if time.Until(idExpiry) > refreshSkew {
				return actor, nil
			}
			// Valid but expiring soon: try to refresh, but keep the working
			// session if refresh fails — don't kill a still-valid token.
			if rt, err := c.Cookie(refreshCookie); err == nil {
				if refreshed, err := s.refresh(c, rt); err == nil {
					return refreshed, nil
				} else {
					log.Warnf("proactive token refresh failed, using still-valid session: %v", err)
				}
			}
			return actor, nil
		}
		// Only an expired token is refreshable; tampered/invalid tokens are not.
		var expiredErr *oidc.TokenExpiredError
		if !errors.As(err, &expiredErr) {
			return Actor{}, err
		}
	}
	// Session token is expired, or its cookie already dropped at Max-Age.
	// Refresh is mandatory here; without a refresh cookie there's nothing to do.
	rt, err := c.Cookie(refreshCookie)
	if err != nil {
		if cookieErr != nil {
			return Actor{}, cookieErr
		}
		return Actor{}, errors.New("session expired and no refresh token available")
	}
	actor, err := s.refresh(c, rt)
	if err != nil {
		s.clearSessionCookies(c)
		return Actor{}, err
	}
	return actor, nil
}

// legacyReadPrefixes are the v1 GET endpoints external automation relies on
// (deployment state only — never repositories/credentialsSecrets/applications
// config, nor any /api/v2/*). Opened to anonymous readers when
// ENVROUTER_ANONYMOUS_LEGACY_READS is set.
var legacyReadPrefixes = []string{
	"/api/v1/refBindings",
	"/api/v1/environments",
	"/api/v1/instances",
	"/api/v1/instancePods",
	"/api/v1/git/refs",
	"/api/v1/git/repositories", // commit lookups (.../commits/:sha)
}

// isLegacyReadPath matches an exact allow-listed path or any sub-path under it
// (the trailing-slash check keeps the boundary tight, so "/api/v1/refBindings"
// won't match a hypothetical "/api/v1/refBindingsX").
func isLegacyReadPath(path string) bool {
	for _, p := range legacyReadPrefixes {
		if path == p || strings.HasPrefix(path, p+"/") {
			return true
		}
	}
	return false
}

// bearerToken extracts the token from an `Authorization: Bearer <token>`
// header, or "" if absent/malformed.
func bearerToken(r *http.Request) string {
	const prefix = "Bearer "
	h := r.Header.Get("Authorization")
	if len(h) > len(prefix) && strings.EqualFold(h[:len(prefix)], prefix) {
		return h[len(prefix):]
	}
	return ""
}

// Middleware guards /api/* when auth is enabled and attaches the Actor to
// the gin context in every mode. Two machine-auth escape hatches apply to
// /api/* before the session gate: a static bearer token (full admin) and an
// optional anonymous-read allow-list for legacy v1 GETs.
func (s *Service) Middleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		isAPI := strings.HasPrefix(c.Request.URL.Path, "/api/")

		// Static API token: full-access machine auth, independent of OIDC.
		// Constant-time compare; wrong/absent token falls through.
		if isAPI && s.cfg.APIToken != "" {
			if tok := bearerToken(c.Request); tok != "" &&
				subtle.ConstantTimeCompare([]byte(tok), []byte(s.cfg.APIToken)) == 1 {
				c.Set(actorKey, Actor{
					UserIdentifier: "api-token",
					FullName:       "API token",
					Admin:          true,
					IP:             c.ClientIP(),
				})
				c.Next()
				return
			}
		}

		if !s.Enabled() || !isAPI {
			c.Set(actorKey, Actor{IP: c.ClientIP()})
			c.Next()
			return
		}

		// Anonymous legacy reads: allow-listed v1 GETs without a session.
		if s.cfg.AnonymousLegacyReads && c.Request.Method == http.MethodGet && isLegacyReadPath(c.Request.URL.Path) {
			c.Set(actorKey, Actor{IP: c.ClientIP()})
			c.Next()
			return
		}

		actor, err := s.actorFromSession(c)
		if err != nil {
			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "session invalid or expired"})
			return
		}
		actor.IP = c.ClientIP()
		// re-check on every request, not just at login: group config may have
		// changed since the session cookie was issued
		if !s.CanView(actor) {
			c.AbortWithStatusJSON(http.StatusForbidden, gin.H{"error": "your account is not in a group permitted to view EnvRouter"})
			return
		}
		c.Set(actorKey, actor)
		c.Next()
	}
}

// ActorFromContext returns the request actor; anonymous (IP only) when auth
// is disabled.
func ActorFromContext(c *gin.Context) Actor {
	if v, ok := c.Get(actorKey); ok {
		if actor, ok := v.(Actor); ok {
			return actor
		}
	}
	return Actor{IP: c.ClientIP()}
}

func randomNonce() (string, error) {
	b := make([]byte, 24)
	if _, err := rand.Read(b); err != nil {
		return "", err
	}
	return base64.RawURLEncoding.EncodeToString(b), nil
}

func (s *Service) LoginHandler(c *gin.Context) {
	if !s.Enabled() {
		c.Redirect(http.StatusFound, "/")
		return
	}
	rd := safeRedirect(c.Query("rd"))
	nonce, err := randomNonce()
	if err != nil {
		c.AbortWithStatus(http.StatusInternalServerError)
		return
	}
	c.SetSameSite(http.SameSiteLaxMode)
	c.SetCookie(stateCookie, nonce+"|"+url.QueryEscape(rd), 600, "/", "", !s.cfg.InsecureCookie, true)
	c.Redirect(http.StatusFound, s.oauth.AuthCodeURL(nonce))
}

func (s *Service) CallbackHandler(c *gin.Context) {
	if !s.Enabled() {
		c.Redirect(http.StatusFound, "/")
		return
	}
	stateValue, err := c.Cookie(stateCookie)
	if err != nil {
		c.String(http.StatusBadRequest, "missing login state; start again at /auth/login")
		return
	}
	nonce, escapedRd, found := strings.Cut(stateValue, "|")
	if !found || c.Query("state") != nonce {
		c.String(http.StatusBadRequest, "login state mismatch; start again at /auth/login")
		return
	}
	unescapedRd, _ := url.QueryUnescape(escapedRd)
	rd := safeRedirect(unescapedRd)
	token, err := s.oauth.Exchange(c.Request.Context(), c.Query("code"))
	if err != nil {
		log.Errorf("OIDC code exchange failed: %v", err)
		c.String(http.StatusBadGateway, "token exchange with the identity provider failed")
		return
	}
	rawIDToken, ok := token.Extra("id_token").(string)
	if !ok {
		c.String(http.StatusBadGateway, "identity provider returned no id_token")
		return
	}
	_, idExpiry, err := s.actorFromIDToken(c.Request.Context(), rawIDToken)
	if err != nil {
		log.Errorf("OIDC id_token rejected: %v", err)
		c.String(http.StatusUnauthorized, "identity token rejected: %v", err)
		return
	}
	// no CanView rejection here: the session is issued even for group-less
	// users so the SPA can show its access-denied screen (userinfo reports
	// canView=false); the API middleware 403s all their data requests
	s.setSessionCookies(c, rawIDToken, idExpiry, token)
	c.SetCookie(stateCookie, "", -1, "/", "", !s.cfg.InsecureCookie, true)
	c.Redirect(http.StatusFound, rd)
}

func (s *Service) LogoutHandler(c *gin.Context) {
	s.clearSessionCookies(c)
	c.Status(http.StatusNoContent)
}

func (s *Service) UserinfoHandler(c *gin.Context) {
	if !s.Enabled() {
		// auth disabled: anonymous actor can do everything
		c.JSON(http.StatusOK, gin.H{
			"enabled":       false,
			"authenticated": false,
			"canDeploy":     s.CanDeploy(Actor{}),
			"canConfigure":  s.CanConfigure(Actor{}),
		})
		return
	}
	if actor, err := s.actorFromSession(c); err == nil {
		c.JSON(http.StatusOK, gin.H{
			"enabled":        true,
			"authenticated":  true,
			"userIdentifier": actor.UserIdentifier,
			"fullName":       actor.FullName,
			"email":          actor.Email,
			"groups":         actor.Groups,
			"canView":        s.CanView(actor),
			"canDeploy":      s.CanDeploy(actor),
			"canConfigure":   s.CanConfigure(actor),
		})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"enabled":       true,
		"authenticated": false,
		"canDeploy":     s.CanDeploy(Actor{}),
		"canConfigure":  s.CanConfigure(Actor{}),
	})
}